// projects/cs-lib/src/lib/types/enum/dateparts.enum.ts
/* P r o p r i e t a r y N o t i c e */
/* Unpublished © 2020 Allscripts Healthcare, LLC and/or its affiliates. All Rights
Reserved.
*
* P r o p r i e t a r y N o t i c e: This software has been provided pursuant to a License Agreement, with Allscripts
Healthcare, LLC and/or its affiliates, containing restrictions on its use. This software contains valuable trade secrets
and proprietary information of Allscripts Healthcare, LLC and/or its affiliates and is protected by trade secret and
copyright law. This software may not be copied or distributed in any form or medium, disclosed to any third parties,
or used in any manner not provided for in said License Agreement except with prior written authorization from
Allscripts Healthcare, LLC and/or its affiliates. Notice to U.S. Government Users: This software is “Commercial
Computer Software.”
Allscripts CareQuality™ is a trademark of Allscripts Healthcare, LLC and/or its affiliates.
*
*
*/
/* P r o p r i e t a r y N o t i c e */

/** Enumeration of access roles */
export enum DateParts {
    Years = 1,
    Months = 2,
    Days = 3,
    Hours = 4,
    Minutes = 5,
}

export enum DateTimePartsInNumber {
    NumberOfMonthInYear = 12,
    NumberOfDaysInMonth = 30,
    NumberOfMiliSecondInDays = 86400000,
    NumberOfMiliSecondInHour = 3600000,
    NumberOfMiliSecondInMinutes = 60000,
}